+++ app/src/main/java/hr/avrbanac/docsis/app/PreEqualizationApplication.java
@@ -3,7 +3,11 @@ package hr.avrbanac.docsis.app;
 import javafx.application.Application;
 import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
+import javafx.scene.image.Image;
 import javafx.stage.Stage;
+
+import java.util.ArrayList;
+import java.util.Objects;
 
 import static hr.avrbanac.docsis.app.PreEqAppConfiguration.*;
 
@@ -14,6 +18,10 @@ public class PreEqualizationApplication extends Application {
         FXMLLoader fxmlLoader = new FXMLLoader(PreEqualizationApplication.class.getResource(APP_VIEW_FILE));
         Scene scene = new Scene(fxmlLoader.load(), APP_WIDTH, APP_HEIGHT);
         scene.getStylesheets().add(APP_CSS_FILE);
+        stage.getIcons().addAll(
+                new Image(Objects.requireNonNull(PreEqualizationApplication.class.getResourceAsStream(ICON_16_FILE))),
+                new Image(Objects.requireNonNull(PreEqualizationApplication.class.getResourceAsStream(ICON_32_FILE))),
+                new Image(Objects.requireNonNull(PreEqualizationApplication.class.getResourceAsStream(ICON_64_FILE))));
         stage.setTitle(APP_TITLE);
         stage.setScene(scene);
         stage.setResizable(false);