+++ lib/src/main/java/hr/avrbanac/docsis/lib/analysis/PreEqAnalysis.java
@@ -32,12 +32,21 @@ public class PreEqAnalysis {
     private final int fftSize;
 
     /**
-     * Elapsed time in nanoseconds needed for FFT preparation and calculation.
+     * Elapsed time in nanoseconds spent in calculation.
      */
-    private long elapsedTime = -1L;
+    private long elapsedTime;
+    /**
+     * Complex array represents In Channel Frequency Response calculated only once, first time required.
+     */
+    private Complex[] fftICFR = null;
+    /**
+     * Double array represents In Channel Frequency Response Magnitude calculated only once, first time required.
+     */
+    private double[] fftICFRMag = null;
 
     public PreEqAnalysis(final PreEqData preEqData) {
         this.preEqData = preEqData;
+        long start = System.nanoTime();
         this.fftSize = MathUtility.getMinFFTSize(
                 preEqData.getTapCount(),
                 preEqData.getMainTapIndex(),
@@ -48,6 +57,138 @@ public class PreEqAnalysis {
         if (fftSize < 0) {
             throw PreEqException.FFT_TAP_COUNT_ERROR;
         }
+
+        elapsedTime = System.nanoTime() - start;
+    }
+
+    /**
+     * Returns the number of input / output points for FFT.
+     * @return int FFT size
+     * @see MathUtility#getMinFFTSize(int, int, PreEqFFTInputFormat, int, int)
+     */
+    public int getFFTSize() {
+        return fftSize;
+    }
+
+    /**
+     * Returns elapsed time in nanoseconds (time needed for FFT preparation and calculation).
+     * @return long elapsed time in ns
+     */
+    public long getElapsedTime() {
+        return elapsedTime;
+    }
+
+    /**
+     * Returns ICFR(In Channel Frequency Response) data as a result of forward FFT (Fast Fourier Transform).
+     * As describe for {@link #prepareComplexFFTInput()}, after forward FFT, rotation needs to be done for the FFT output.
+     * Returned array is an array of complex points. For plotting of the ICFR use {@link #getInChannelFrequencyResponseMagnitude()} method
+     * instead.
+     * @return {@link Complex} array in frequency domain
+     */
+    public Complex[] getInChannelFrequencyResponse() {
+        if (fftICFR != null) return fftICFR;
+
+        long start = System.nanoTime();
+        FastFourierTransformer transformer = new FastFourierTransformer(DFT_NORMALIZATION);
+        fftICFR = createRotatedFFTArray(transformer.transform(prepareComplexFFTInput(), TRANSFORM_TYPE));
+        elapsedTime += System.nanoTime() - start;
+
+        return fftICFR;
+    }
+
+    /**
+     * Returns ICFR (In Channel Frequency Response) data as a result of forward FFT (Fast Fourier Transform).
+     * Implicitly, this method will call {@link #getInChannelFrequencyResponse()} and convert values to magnitude values for plotting.
+     * @return array of double values representing magnitudes of FFT output complex points
+     */
+    public double[] getInChannelFrequencyResponseMagnitude() {
+        if (fftICFRMag != null) return fftICFRMag;
+
+        Complex[] fftResult = getInChannelFrequencyResponse();
+        fftICFRMag = new double[fftSize];
+
+        long start = System.nanoTime();
+        for (int i = 0; i < fftSize; i++) {
+            fftICFRMag[i] = 20 * Math.log10(Math.hypot(fftResult[i].getReal(), fftResult[i].getImaginary()));
+        }
+
+        elapsedTime += System.nanoTime() - start;
+        return fftICFRMag;
+    }
+
+    /**
+     * Time domain reflectometry calculation method.
+     * <p>
+     * When a propagating wave hits the impedance mismatch point, a micro-reflection (MR) occurs. This MR is reflected towards CM (wave
+     * origin) until it hits another reflection point where it is reflected again towards target (US direction). As a result,
+     * a superposition of the original wave and attenuated reflected one is perceived by the CMTS somewhere in the upstream. However, CMTS
+     * will register reflected wave as a delayed wave in correlation to the original one. This time dilatation can be used together with the
+     * known channel frequency and speed of the wave propagation to calculate the distance between two reflection points.
+     * </p>
+     * <p>
+     * When addressing these MRs, there is a distinction: near and far reflections - assuming that the first n energy taps after the main
+     * one are considered near to the CM device if the level of the MR is above certain threshold.
+     * </p>
+     * <p>
+     * Calculation will be carried out with the highest post-main tap using hyperbolic interpolation. For this interpolation process, 3
+     * points will be used: the highest MR tap, closest left and closest right tap value. Exception to this rule is calculating
+     * interpolation for first post-main tap since we cannot use previous one as left point (main tap is maxed out). Instead, last pre-main
+     * tap will be used with according x-axis value (2 places offset instead of just one). Similarly, last post-main tap has no right point,
+     * so imaginary right point with the same value as that to the right of the last one will be used, effectively fixing calculation to the
+     * middle of the last tap. This will lower the precision of calculating interpolation for the last energy tap. (This is the last tap and
+     * the time dilatation is the longest, so lowering the precision is not so big of a problem.)
+     * </p>
+     *
+     * @param channelWidth {@link ChannelWidth} carrying the information about width of the channel and symbol rate
+     * @param nearPostMainTapCount int count of the post-main energy taps considered near reflections
+     * @param onlyFarReflections boolean value - true if near reflections should be left out of the calculation
+     * @return double TDR value representing distance between the strongest MR point and first strong downstream reflection point
+     */
+    public double getTDR(
+            final ChannelWidth channelWidth,
+            final int nearPostMainTapCount,
+            final boolean onlyFarReflections) {
+
+        long start = System.nanoTime();
+        int mainTapIndex = preEqData.getMainTapIndex();
+        int tapCount = preEqData.getTapCount();
+        double maxReflection = preEqData.getTapEnergyRatioBoundary();
+        List<Coefficient> coefficients = preEqData.getCoefficients();
+        long lMTNE = preEqData.getMTNE();
+        int ptr = onlyFarReflections ? mainTapIndex + nearPostMainTapCount : mainTapIndex;
+        int maxTapPtr = 0;
+
+        while (ptr < tapCount) {
+            double currTapEnergyRatio = coefficients.get(ptr).getEnergyRatio(lMTNE);
+            if (currTapEnergyRatio > maxReflection) {
+                maxReflection = currTapEnergyRatio;
+                maxTapPtr = ptr;
+            }
+            ptr++;
+        }
+
+        if (maxTapPtr <= mainTapIndex - 1) throw PreEqException.TDR_CALCULATION_ERROR;
+
+        Complex left = getLeftInterpolationPoint(coefficients, maxTapPtr, mainTapIndex - 1);
+        Complex middle = new Complex(maxTapPtr + 1d, coefficients.get(maxTapPtr).getImag());
+        Complex right = getRightInterpolationPoint(coefficients, maxTapPtr, tapCount);
+        Complex interpolated = MathUtility.calculateParabolicInterpolation(left, middle, right);
+        double result = (interpolated.getReal() - mainTapIndex) * MathUtility.getTDRSpeedFactor(channelWidth.getSymRate());
+        elapsedTime += System.nanoTime() - start;
+
+        return result;
+    }
+
+    public Signature getSignature() {
+        return getSignature(MicroReflectionSeverityThreshold.CABLE_LABS);
+    }
+
+    public Signature getSignature(final MicroReflectionSeverityThreshold thresholdLevel) {
+        long start = System.nanoTime();
+        Signature signature = new Signature(getInChannelFrequencyResponseMagnitude(), thresholdLevel);
+        elapsedTime += System.nanoTime() - start;
+
+        return signature;
     }
 
     /**
@@ -110,117 +251,6 @@ public class PreEqAnalysis {
     }
 
     /**
-     * Returns the number of input / output points for FFT.
-     * @return int FFT size
-     * @see MathUtility#getMinFFTSize(int, int, PreEqFFTInputFormat, int, int)
-     */
-    public int getFFTSize() {
-        return fftSize;
-    }
-
-    /**
-     * Returns elapsed time in nanoseconds (time needed for FFT preparation and calculation).
-     * @return long elapsed time in ns
-     */
-    public long getElapsedTime() {
-        return elapsedTime;
-    }
-
-    /**
-     * Returns ICFR(In Channel Frequency Response) data as a result of forward FFT (Fast Fourier Transform).
-     * As describe for {@link #prepareComplexFFTInput()}, after forward FFT, rotation needs to be done for the FFT output.
-     * Returned array is an array of complex points. For plotting of the ICFR use {@link #getInChannelFrequencyResponseMagnitude()} method
-     * instead.
-     * @return {@link Complex} array in frequency domain
-     */
-    public Complex[] getInChannelFrequencyResponse() {
-        long start = System.nanoTime();
-
-        FastFourierTransformer transformer = new FastFourierTransformer(DFT_NORMALIZATION);
-        Complex[] fftResult = createRotatedFFTArray(transformer.transform(prepareComplexFFTInput(), TRANSFORM_TYPE));
-
-        elapsedTime = System.nanoTime() - start;
-
-        return fftResult;
-    }
-
-    /**
-     * Returns ICFR (In Channel Frequency Response) data as a result of forward FFT (Fast Fourier Transform).
-     * Implicitly, this method will call {@link #getInChannelFrequencyResponse()} and convert values to magnitude values for plotting.
-     * @return array of double values representing magnitudes of FFT output complex points
-     */
-    public double[] getInChannelFrequencyResponseMagnitude() {
-        Complex[] fftResult = getInChannelFrequencyResponse();
-        double[] result = new double[fftSize];
-
-        for (int i = 0; i < fftSize; i++) {
-            result[i] = 20 * Math.log10(Math.hypot(fftResult[i].getReal(), fftResult[i].getImaginary()));
-        }
-
-        return result;
-    }
-
-    /**
-     * Time domain reflectometry calculation method.
-     * <p>
-     * When a propagating wave hits the impedance mismatch point, a micro-reflection (MR) occurs. This MR is reflected towards CM (wave
-     * origin) until it hits another reflection point where it is reflected again towards target (US direction). As a result,
-     * a superposition of the original wave and attenuated reflected one is perceived by the CMTS somewhere in the upstream. However, CMTS
-     * will register reflected wave as a delayed wave in correlation to the original one. This time dilatation can be used together with the
-     * known channel frequency and speed of the wave propagation to calculate the distance between two reflection points.
-     * </p>
-     * <p>
-     * When addressing these MRs, there is a distinction: near and far reflections - assuming that the first n energy taps after the main
-     * one are considered near to the CM device if the level of the MR is above certain threshold.
-     * </p>
-     * <p>
-     * Calculation will be carried out with the highest post-main tap using hyperbolic interpolation. For this interpolation process, 3
-     * points will be used: the highest MR tap, closest left and closest right tap value. Exception to this rule is calculating
-     * interpolation for first post-main tap since we cannot use previous one as left point (main tap is maxed out). Instead, last pre-main
-     * tap will be used with according x-axis value (2 places offset instead of just one). Similarly, last post-main tap has no right point,
-     * so imaginary right point with the same value as that to the right of the last one will be used, effectively fixing calculation to the
-     * middle of the last tap. This will lower the precision of calculating interpolation for the last energy tap. (This is the last tap and
-     * the time dilatation is the longest, so lowering the precision is not so big of a problem.)
-     * </p>
-     *
-     * @param channelWidth {@link ChannelWidth} carrying the information about width of the channel and symbol rate
-     * @param nearPostMainTapCount int count of the post-main energy taps considered near reflections
-     * @param onlyFarReflections boolean value - true if near reflections should be left out of the calculation
-     * @return double TDR value representing distance between the strongest MR point and first strong downstream reflection point
-     */
-    public double getTDR(
-            final ChannelWidth channelWidth,
-            final int nearPostMainTapCount,
-            final boolean onlyFarReflections) {
-
-        int mainTapIndex = preEqData.getMainTapIndex();
-        int tapCount = preEqData.getTapCount();
-        double maxReflection = preEqData.getTapEnergyRatioBoundary();
-        List<Coefficient> coefficients = preEqData.getCoefficients();
-        long lMTNE = preEqData.getMTNE();
-        int ptr = onlyFarReflections ? mainTapIndex + nearPostMainTapCount : mainTapIndex;
-        int maxTapPtr = 0;
-
-        while (ptr < tapCount) {
-            double currTapEnergyRatio = coefficients.get(ptr).getEnergyRatio(lMTNE);
-            if (currTapEnergyRatio > maxReflection) {
-                maxReflection = currTapEnergyRatio;
-                maxTapPtr = ptr;
-            }
-            ptr++;
-        }
-
-        if (maxTapPtr <= mainTapIndex - 1) throw PreEqException.TDR_CALCULATION_ERROR;
-
-        Complex left = getLeftInterpolationPoint(coefficients, maxTapPtr, mainTapIndex - 1);
-        Complex middle = new Complex(maxTapPtr + 1d, coefficients.get(maxTapPtr).getImag());
-        Complex right = getRightInterpolationPoint(coefficients, maxTapPtr, tapCount);
-        Complex interpolated = MathUtility.calculateParabolicInterpolation(left, middle, right);
-
-        return (interpolated.getReal() - mainTapIndex) * MathUtility.getTDRSpeedFactor(channelWidth.getSymRate());
-    }
-
-    /**
      * Helper method will return {@link Complex} wrapper where real value is a number of the tap (not an array index) and imaginary value is
      * the energy ratio of the coefficient left of the max reflection tap. There is a special case, when the max reflection tap is the first
      * one after the main tap. Then the calculation is done with the one left of the main tap taking into account it's fixed position.
@@ -259,7 +289,4 @@ public class PreEqAnalysis {
                 ? new Complex(middlePtr + 2d, coefficients.get(middlePtr - 1).getImag())
                 : new Complex(middlePtr + 2d, coefficients.get(middlePtr + 1).getImag());
     }
-
-
-
 }